Improve usability.
Add OddValidator for checking patch radius.

Add QIntValidators on PatchInfoWidget to make sure the user cannot move the patches
outside of the image.

Add QIntValidators on TopPatchesWidget to make sure reasonable values are input.

Recompute the projection matrix if the patch size is changed. Otherwise Eigen throws
an exception that the ProjectionMatrix dimensions do not allow mutliplication by
vectorized patches of the new size.